+++ src/main/java/autolotto/model/ProfitMatcher.java
@@ -5,12 +5,12 @@ import java.util.Map;
 
 public class ProfitMatcher {
 
-    private static final Map<Integer, Float>  matchProfitList = new HashMap<Integer, Float>(){
+    private static final Map<Integer, Integer>  matchProfitList = new HashMap<Integer, Integer>(){
         {
-            put(3,5000.0f);
-            put(4,50000.0f);
-            put(5,1500000.0f);
-            put(6,2000000000.0f);
+            put(3,5000);
+            put(4,50000);
+            put(5,1500000);
+            put(6,2000000000);
         }
     };
